refactor: relocate npc + agent_backend out of core/ to top-level
Public-facing protocols pack authors implement against don't belong
under ``openrange.core`` — that namespace is for internal building
blocks (manifest parsing, episode service internals, runtime
backings, snapshot store, graph machinery). Pack authors should
import from ``openrange`` (or top-level submodules), never from
``.core``.

Moves:
* ``openrange.core.agent_backend`` -> ``openrange.agent_backend``
* ``openrange.core.npc`` -> ``openrange.npc``

Both now sit at the same tier as ``openrange.llm`` and
``openrange.runtime`` — public, top-level, owned by users and pack
authors.

Top-level ``openrange`` re-exports the NPC family
(``NPC``, ``AgentNPC``, ``NPCRegistry``, ``NPCError``, ``NPCS``)
alongside the existing ``LLMBackend`` / ``AgentBackend`` exports,
so the typical pack import becomes ``from openrange import NPC,
AgentNPC, ...`` — no submodule reach-through.

The cyber_webapp pack imports were updated to match.
