+++ packs/cyber_webapp/cyber_webapp/npcs/browsing_user.py
@@ -15,7 +15,7 @@ from __future__ import annotations
 from collections.abc import Mapping
 
 from cyber_webapp.npcs import _HTTPCadenceNPC
-from openrange.core.npc import NPC
+from openrange.npc import NPC
 
 _DEFAULT_PATHS: tuple[str, ...] = ("/",)
